+++ packages/cli/__tests__/zosuss/__unit__/issue/ssh/Ssh.handler.unit.test.ts
@@ -11,7 +11,7 @@
 
 jest.mock("../../../../../../zosuss/lib/Shell");
 
-import { IHandlerParameters, IProfile, CommandProfiles } from "@zowe/imperative";
+import { IHandlerParameters, IProfile, CommandProfiles, ConnectionPropsForSessCfg } from "@zowe/imperative";
 import * as SshHandler from "../../../../../src/zosuss/issue/ssh/Ssh.handler";
 import * as SshDefinition from "../../../../../src/zosuss/issue/ssh/Ssh.definition";
 import { Shell } from "@zowe/zos-uss-for-zowe-sdk";
@@ -33,6 +33,13 @@ const UNIT_TEST_SSH_PROF_OPTS_PRIVATE_KEY = {
     user: "someone",
     privateKey: normalize(join(__dirname, "..", "..", "..", "..", "..", "..", "zosuss", "__tests__", "__unit__", "__resources__", "fake_id_rsa"))
 };
+const UNIT_TEST_SSH_PROF_OPTS_PRIVATE_KEY_WITH_PASSPHRASE = {
+    host: "somewhere.com",
+    port: "22",
+    user: "someone",
+    privateKey: normalize(join(__dirname, "..", "..", "..", "..", "..", "..", "zosuss", "__tests__", "__unit__", "__resources__", "fake_id_rsa")),
+    keyPassPhrase: "dummyPassPhrase123"
+};
 
 // A mocked profile map with ssh profile
 const UNIT_TEST_PROFILE_MAP = new Map<string, IProfile[]>();
@@ -53,8 +60,17 @@ UNIT_TEST_PROFILE_MAP_PRIVATE_KEY.set(
         ...UNIT_TEST_SSH_PROF_OPTS_PRIVATE_KEY
     }]
 );
-const UNIT_TEST_PROFILES_SSH_PRIVATE_KEY = new CommandProfiles(UNIT_TEST_PROFILE_MAP_PRIVATE_KEY);
+const UNIT_TEST_PROFILE_MAP_PRIVATE_KEY_WITH_PASSPHRASE = new Map<string, IProfile[]>();
+UNIT_TEST_PROFILE_MAP_PRIVATE_KEY_WITH_PASSPHRASE.set(
+    "ssh", [{
+        name: "ssh",
+        type: "ssh",
+        ...UNIT_TEST_SSH_PROF_OPTS_PRIVATE_KEY_WITH_PASSPHRASE
+    }]
+);
 
+const UNIT_TEST_PROFILES_SSH_PRIVATE_KEY = new CommandProfiles(UNIT_TEST_PROFILE_MAP_PRIVATE_KEY);
+const UNIT_TEST_PROFILES_SSH_PRIVATE_KEY_WITH_PASSPHRASE = new CommandProfiles(UNIT_TEST_PROFILE_MAP_PRIVATE_KEY_WITH_PASSPHRASE);
 // Mocked parameters for the unit tests
 const DEFAULT_PARAMETERS: IHandlerParameters = mockHandlerParameters({
     arguments: UNIT_TEST_SSH_PROF_OPTS,
@@ -68,6 +84,13 @@ const DEFAULT_PARAMETERS_PRIVATE_KEY: IHandlerParameters = mockHandlerParameters
     positionals: ["zos-uss", "issue", "ssh"],
     definition: SshDefinition.SshDefinition,
     profiles: UNIT_TEST_PROFILES_SSH_PRIVATE_KEY
+});
+
+const DEFAULT_PARAMETERS_KEY_PASSPHRASE: IHandlerParameters = mockHandlerParameters({
+    arguments: UNIT_TEST_SSH_PROF_OPTS_PRIVATE_KEY_WITH_PASSPHRASE,
+    positionals: ["zos-uss", "issue", "ssh"],
+    definition: SshDefinition.SshDefinition,
+    profiles: UNIT_TEST_PROFILES_SSH_PRIVATE_KEY_WITH_PASSPHRASE,
 });
 
 const testOutput = "TEST OUTPUT";
@@ -90,6 +113,62 @@ describe("issue ssh handler tests", () => {
         expect(testOutput).toMatchSnapshot();
     });
 
+    it("should be able to get stdout with private key and key passphrase", async () => {
+        Shell.executeSsh = jest.fn(async (session, command, stdoutHandler) => {
+            stdoutHandler(testOutput);
+        });
+        const handler = new SshHandler.default();
+        const params = Object.assign({}, ...[DEFAULT_PARAMETERS_KEY_PASSPHRASE]);
+        params.arguments.command = "echo test";
+        await handler.process(params);
+        expect(Shell.executeSsh).toHaveBeenCalledTimes(1);
+        expect(testOutput).toMatchSnapshot();
+    });
+    it("should prompt user for keyPassphrase if none is stored and privateKey requires one", async () => {
+        Shell.executeSsh = jest.fn(async (session, command, stdoutHandler) => {
+            stdoutHandler(testOutput);
+        });
+        const handler = new SshHandler.default();
+        const params = Object.assign({}, ...[DEFAULT_PARAMETERS_KEY_PASSPHRASE]);
+        params.arguments.command = "echo test";
+        jest.spyOn(handler,"processCmd").mockImplementationOnce(() => {throw new Error("but no passphrase given");});
+        jest.spyOn(ConnectionPropsForSessCfg as any,"getValuesBack").mockReturnValue(() => ({
+            keyPassphrase: "validPassword"
+        }));
+        await handler.process(params);
+        expect(Shell.executeSsh).toHaveBeenCalledTimes(1);
+        expect(testOutput).toMatchSnapshot();
+    });
+    it("should reprompt user for keyPassphrase up to 3 times if stored passphrase failed", async () => {
+        Shell.executeSsh = jest.fn(async (session, command, stdoutHandler) => {
+            stdoutHandler(testOutput);
+        });
+        const handler = new SshHandler.default();
+        const params = Object.assign({}, ...[DEFAULT_PARAMETERS_KEY_PASSPHRASE]);
+        params.arguments.command = "echo test";
+        jest.spyOn(handler,"processCmd").mockImplementationOnce(() => {throw new Error("bad passphrase?");});
+        jest.spyOn(ConnectionPropsForSessCfg as any,"getValuesBack").mockReturnValue(() => ({
+            keyPassphrase: "validPassword"
+        }));
+        await handler.process(params);
+        expect(Shell.executeSsh).toHaveBeenCalledTimes(1);
+        expect(testOutput).toMatchSnapshot();
+    });
+    it("should fail if user fails to enter incorrect key passphrase in 3 attempts", async () => {
+        const testOutput = "Maximum retry attempts reached. Authentication failed.";
+        Shell.executeSsh = jest.fn(async (session, command, stdoutHandler) => {
+            stdoutHandler(testOutput);
+        });
+        const handler = new SshHandler.default();
+        const params = { ...DEFAULT_PARAMETERS_KEY_PASSPHRASE };
+        params.arguments.command = "echo test";
+        jest.spyOn(handler, "processCmd").mockImplementation(() => {
+            throw new Error("bad passphrase?");
+        });
+        await expect(handler.process(params)).rejects.toThrow("Maximum retry attempts reached. Authentication failed.");
+        expect(handler.processCmd).toHaveBeenCalledTimes(4);
+        expect(testOutput).toMatchSnapshot();
+    });
     it("should be able to get stdout with privateKey", async () => {
         Shell.executeSsh = jest.fn(async (session, command, stdoutHandler) => {
             stdoutHandler(testOutput);
@@ -101,7 +180,6 @@ describe("issue ssh handler tests", () => {
         expect(Shell.executeSsh).toHaveBeenCalledTimes(1);
         expect(testOutput).toMatchSnapshot();
     });
-
     it("should be able to get stdout with cwd option", async () => {
         Shell.executeSshCwd = jest.fn(async (session, command, cwd, stdoutHandler) => {
             stdoutHandler(testOutput);
@@ -114,5 +192,4 @@ describe("issue ssh handler tests", () => {
         expect(Shell.executeSshCwd).toHaveBeenCalledTimes(1);
         expect(testOutput).toMatchSnapshot();
     });
-
 });